Add tests
Tests are good. I can't believe I hadn't already added them. In addition
to adding the tests, one change needs to be made to `S3`. The `Session`
instance needs to be needs to be available externally so that placebo
can intercept traffic to AWS. Fortunately creating a `Session` instance
results in no network traffic so it's safe to instantiate it in
`init_app`.