Add a method to check for the existence of a key
Currently the only way for a user of Henson-S3 to know if a key already
exists on S3 is to download the file. A new coroutine (`check`) is being
added that will check for the key using a `HEAD`.

In addition to the code check, tests and the placebo responses needed
for them are being added.