Add the public API to the documentation
In order for people to know how to use `S3`, it's important to include
its API in Henson-S3's documentation. In order to make the documentation
a little easier to read and keep up-to-date, the intersphinx mappings
are being updated and some of the docstrings are being updated to better
utilize `typing`.