Add Python 3.6 to the supported versions
Along with listing 3.6 in `setup.py`, tox will now use it as its default
version of Python. The Travis CI configuration is being updated to use a
build matrix since 3.5 and 3.6 don't play nicely together there (the
config is also being simplified a bit to remove some of the noise).