Stop updating the copyright year
I don't want to have to remember to update the copyright year each time
I make a change in a new calendar year. (Obviously I failed to do that
already this year.) By removing the year completely from the LICENSE
file and letting Python set it in the documentation, I shouldn't ever
have to think about it again.